requirements.txt
# Core dependencies
numpy>=1.24.0
scipy>=1.10.0

# Document processing
python-frontmatter>=1.0.0
markdown>=3.4.0
beautifulsoup4>=4.12.0

# Search and embeddings
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
rank-bm25>=0.2.2

# Clustering
scikit-learn>=1.3.0
hdbscan>=0.8.33
umap-learn>=0.5.3

# LLM integration
openai>=1.0.0
tiktoken>=0.5.0
requests>=2.31.0  # For DeepSeek API calls

# Web UI
Flask>=3.0.0
Flask-CORS>=4.0.0

# CLI and utilities
click>=8.1.0
rich>=13.0.0
tqdm>=4.65.0
pyyaml>=6.0

# Optional: Local LLM support
# ollama>=0.1.0

# Development
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
